Add Glean overview article (#276)
Add a first pass on a Glean overview article, so we have a single place to point to.

The goal is to have a landing page for Engineers from other teams, Engineering Managers, Product Managers, etc. This means giving them a basic overview of what Glean is and where to find more information.