Update definitions; remove MAU definition
Previously, the segments were defined with 28 days of historical data. jklukas strongly recommended we use only 27 days of history, so that the segment could be implemented as a simple performant view.

As per the internal gdoc, the best way to implement this is to make the most recent week a 6 day week and require only >=1 5 URI day for that week. It makes the segment definitions more complicated to explain, but the specific definitions are not what make these segments useful, so this is likely a net positive change.

Also I removed the definition about MAU because it looks like there will be a potentially-long running debate around it, and we don't need to block everything else on resolving that discussion.