Add a note about GLAM access being limited (#645)
Currently GLAM access is limited to people with LDAP access, this might change in the future but let's
at least make this more clear by putting a note about it right on top (rather than beneath the fold). Link 
to a bug I just filed so people can track the current status.