Update markdown-link-check
This also includes dependency updates, avoiding a vulnerable version of
marked and json-schema.
Note that markdown-spellcheck still depends on the vulnerable marked and
has no updates.
That's okayish. We only run that tooling in CI, its docs only here and
its an internal dependency. So unless someone actually puts in the
wrong markdown we're fine.