Helps clarify example label in sql guidelines
The example for `Do not include multiple arguments on one line` has a second bad example that may confuse the reader.

The reason is it fixes a problem the bad example has but also introduces a new problem the bad example didn't have. This change would move both issues into the first bad example and fix them both in the good example.